+++ src/NetTopologySuite.IO.GPX/GpxReader.cs
@@ -132,19 +132,31 @@ namespace NetTopologySuite.IO
                     break;
                 }
 
-                bool foundMetadata = false;
+                bool expectingMetadata = true;
                 bool expectingExtensions = true;
                 do
                 {
+                    if (expectingMetadata)
+                    {
+                        expectingMetadata = false;
+                        if (reader.Name == "metadata")
+                        {
+                            ReadMetadata(reader, settings, creator, visitor);
+                            if (!ReadTo(reader, XmlNodeType.Element, XmlNodeType.EndElement))
+                            {
+                                break;
+                            }
+                        }
+                        else
+                        {
+                            visitor.VisitMetadata(new GpxMetadata(creator));
+                        }
+                    }
+
                     switch (reader.Name)
                     {
                         // ideally, it should all be in this order, since the XSD validation
                         // would fail otherwise, but whatever.
-                        case "metadata":
-                            ReadMetadata(reader, settings, creator, visitor);
-                            foundMetadata = true;
-                            break;
-
                         case "wpt":
                             ReadWaypoint(reader, settings, visitor);
                             break;
@@ -167,27 +179,16 @@ namespace NetTopologySuite.IO
                             }
 
                             break;
-                        case "":
-                        case "gpx":
+
+                        case string _ when settings.IgnoreUnexpectedChildrenOfTopLevelElement:
+                            reader.Skip();
                             break;
+
                         default:
-                            if (settings.IgnoreBadElements)
-                            {
-                                ReadTo(reader, XmlNodeType.EndElement, XmlNodeType.EndElement);
-                            }
-                            else
-                            {
-                                throw new XmlException($"Invalid xml node '{reader.Name}'");
-                            }
-                            break;
+                            throw new XmlException($"Unexpected xml node '{reader.Name}'");
                     }
                 }
                 while (ReadTo(reader, XmlNodeType.Element, XmlNodeType.EndElement));
-
-                if (!foundMetadata)
-                {
-                    visitor.VisitMetadata(new GpxMetadata(creator));
-                }
             }
         }